5002 - Исправлены замечания. Изменил тип id на long. Переставил Autowired на поле. Добавил метод editPupil. Заменил getById() на findById()

+++ src/main/java/com/kataacademy/schoolportal/common/repository/persons/PupilRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PupilRepository extends JpaRepository<Pupil, Integer> {
+public interface PupilRepository extends JpaRepository<Pupil, Long> {
 
 }
